requirements.txt
Flask
Flask-SQLAlchemy
Flask-Session
psycopg2-binary
SQLAlchemy
requests
pytest
sqlformat
python-dotenv
